// src/generated/models/Pm_fiscalperiodsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_fiscalperiodsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_fiscalperiodsstatecode = keyof typeof Pm_fiscalperiodsstatecode;
export const Pm_fiscalperiodsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_fiscalperiodsstatuscode = keyof typeof Pm_fiscalperiodsstatuscode;

export interface Pm_fiscalperiodsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_enddate?: string;
  pm_fiscalperiodid: string;
  pm_fiscalyear?: number;
  pm_isclosed?: boolean;
  pm_iscurrentperiod?: boolean;
  pm_periodname?: string;
  pm_periodnumber?: number;
  pm_startdate?: string;
  statecode: Pm_fiscalperiodsstatecode;
  statuscode?: Pm_fiscalperiodsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_fiscalperiods extends Pm_fiscalperiodsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_isclosedname?: string;
  pm_iscurrentperiodname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}